pyproject.toml
[tool.poetry]
name = "swap-faces-api"
version = "0.1.0"
description = ""
authors = ["leandro-driguez <leandro_driguez@outlook.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "swap_faces_api"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12" 
fastapi = "0.101.1"
numpy = "1.23.5"
pillow = "10.0.0"
uvicorn = "0.23.2"
python-multipart = "^0.0.6"
pydantic = "^2.3.0"
annotated-types = "^0.5.0"
anyio = "^4.0.0"
click = "^8.1.7"
colorama = "^0.4.6"
exceptiongroup = "^1.1.3"
h11 = "^0.14.0"
idna = "^3.4"
pydantic-core = "2.6.3"
sniffio = "^1.3.0"
starlette = ">=0.27.0,<0.28.0"
typing-extensions = "^4.7.1"
opencv-python = "^4.8.0.76"
onnx = "^1.14.1"
onnxruntime = "^1.15.1"
insightface = "^0.7.3"
psutil = "^5.9.5"
tk = "^0.1.0"
customtkinter = "^5.2.0"
tkinterdnd2 = "^0.3.0"
tkinterdnd2-universal = "^1.7.3"
tensorflow = "^2.13.0"
opennsfw2 = "^0.11.0"
protobuf = "^4.24.2"
tqdm = "^4.66.1"
gfpgan = "^1.3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
